+++ find.js
@@ -7,16 +7,16 @@ CollectionHooks.defineAdvice("find", function (userId, _super, aspects, getTrans
   // args[1] : options
 
   // before
-  _.each(aspects.before, function (aspect) {
-    var r = aspect.call(ctx, userId, args[0], args[1]);
+  _.each(aspects.before, function (o) {
+    var r = o.aspect.call(ctx, userId, args[0], args[1]);
     if (r === false) abort = true;
   });
 
   if (abort) return false;
 
   function after(cursor) {
-    _.each(aspects.after, function (aspect) {
-      aspect.call(ctx, userId, args[0], args[1], cursor);
+    _.each(aspects.after, function (o) {
+      o.aspect.call(ctx, userId, args[0], args[1], cursor);
     });
   }