+++ update.js
@@ -24,18 +24,20 @@ CollectionHooks.defineAdvice("update", function (userId, _super, aspects, getTra
 
   // copy originals for convenience for the "after" pointcut
   if (aspects.after) {
-    prev.mutator = EJSON.clone(args[1]);
-    prev.options = EJSON.clone(args[2]);
-    prev.docs = {};
-    _.each(docs, function (doc) {
-      prev.docs[doc._id] = EJSON.clone(doc);
-    });
+    if (_.some(aspects.after, function (o) { return o.options.fetchPrevious !== false; })) {
+      prev.mutator = EJSON.clone(args[1]);
+      prev.options = EJSON.clone(args[2]);
+      prev.docs = {};
+      _.each(docs, function (doc) {
+        prev.docs[doc._id] = EJSON.clone(doc);
+      });
+    }
   }
 
   // before
-  _.each(aspects.before, function (aspect) {
+  _.each(aspects.before, function (o) {
     _.each(docs, function (doc) {
-      var r = aspect.call(_.extend({transform: getTransform(doc)}, ctx), userId, doc, fields, args[1], args[2]);
+      var r = o.aspect.call(_.extend({transform: getTransform(doc)}, ctx), userId, doc, fields, args[1], args[2]);
       if (r === false) abort = true;
     });
   });
@@ -46,11 +48,11 @@ CollectionHooks.defineAdvice("update", function (userId, _super, aspects, getTra
     var fields = getFields(args[1]);
     var docs = CollectionHooks.getDocs.call(self, collection, {_id: {$in: docIds}}, args[2]).fetch();
 
-    _.each(aspects.after, function (aspect) {
+    _.each(aspects.after, function (o) {
       _.each(docs, function (doc) {
-        aspect.call(_.extend({
+        o.aspect.call(_.extend({
           transform: getTransform(doc),
-          previous: prev.docs[doc._id],
+          previous: prev.docs && prev.docs[doc._id],
           affected: affected,
           err: err
         }, ctx), userId, doc, fields, prev.mutator, prev.options);